+++ src/debug.rs
@@ -33,9 +33,12 @@ impl Plugin for DebugPlugin {
     #[rustfmt::skip]
     fn build(&self, app: &mut App) {
         if cfg!(debug_assertions) {
-            app.add_plugin(WorldInspectorPlugin::new())
-                // .add_plugin(FilterQueryInspectorPlugin::<With<TextureAtlasSprite>>::default())
-                // .add_plugin(ResourceInspectorPlugin::<Time>::default())
+            app
+                .add_plugins((
+                    WorldInspectorPlugin::new(),
+                    // ResourceInspectorPlugin::<Time>::default(),
+                    // FilterQueryInspectorPlugin::<With<TextureAtlasSprite>>::default(),
+                ))
                
                 /* -------------------------------------------------------------------------- */
                 /*                          --- Global Structure ---                          */
@@ -93,9 +96,11 @@ impl Plugin for DebugPlugin {
                 .register_type::<LastTurnActionHistory>()
                 .register_type::<ActionsLogs>()
 
-                .add_plugin(ResourceInspectorPlugin::<ActionHistory>::default())
-                .add_plugin(ResourceInspectorPlugin::<LastTurnActionHistory>::default())
-                .add_plugin(ResourceInspectorPlugin::<ActionsLogs>::default())
+                .add_plugins((
+                    ResourceInspectorPlugin::<ActionHistory>::default(),
+                    ResourceInspectorPlugin::<LastTurnActionHistory>::default(),
+                    ResourceInspectorPlugin::<ActionsLogs>::default(),
+                ))
                 ;
         }
     }