+++ src/ui/combat_panel.rs
@@ -15,7 +15,7 @@ use crate::{
     },
     constants::{
         combat::FIRST_ENEMY_ID,
-        ui::{dialogs::*, style::*},
+        ui::{dialogs::*, style::*, *},
     },
     ui::{
         combat_system::{HpMeter, MpMeter},
@@ -355,7 +355,8 @@ pub fn hud_wall_setup(
         .spawn((
             ImageBundle {
                 style: Style {
-                    size: Size::all(Val::Px(50.)),
+                    width: Val::Px(50.),
+                    height: Val::Px(50.),
                     align_self: AlignSelf::Center,
                     ..default()
                 },
@@ -371,7 +372,7 @@ pub fn hud_wall_setup(
         .spawn((
             NodeBundle {
                 style: Style {
-                    // size: Size::height(Val::Percent(42.)),
+                    // height: Val::Percent(42.),
                     flex_direction: FlexDirection::Row,
                     ..default()
                 },
@@ -412,7 +413,7 @@ pub fn hud_wall_setup(
         .spawn((
             NodeBundle {
                 style: Style {
-                    // size: Size::height(Val::Percent(42.)),
+                    // height: Val::Percent(42.),
                     flex_direction: FlexDirection::Row,
                     ..default()
                 },
@@ -428,14 +429,14 @@ pub fn hud_wall_setup(
                     .spawn((
                         ButtonBundle {
                             style: Style {
-                                size: Size::new(Val::Px(150.0), Val::Px(65.0)),
+                                width: Val::Px(150.0),
+                                height: Val::Px(65.0),
                                 // center button
                                 margin: UiRect::all(Val::Auto),
                                 // horizontally center child text
                                 justify_content: JustifyContent::Center,
                                 // vertically center child text
                                 align_items: AlignItems::Center,
-                                position: UiRect::default(),
                                 ..default()
                             },
                             background_color: NORMAL_BUTTON.into(),
@@ -463,7 +464,7 @@ pub fn hud_wall_setup(
         .spawn((
             NodeBundle {
                 style: Style {
-                    // size: Size::height(Val::Percent(42.)),
+                    // height: Val::Percent(42.),
                     flex_direction: FlexDirection::Row,
                     ..default()
                 },
@@ -479,14 +480,14 @@ pub fn hud_wall_setup(
                     .spawn((
                         ButtonBundle {
                             style: Style {
-                                size: Size::new(Val::Px(150.0), Val::Px(65.0)),
+                                width: Val::Px(150.0),
+                                height: Val::Px(65.0),
                                 // center button
                                 margin: UiRect::all(Val::Auto),
                                 // horizontally center child text
                                 justify_content: JustifyContent::Center,
                                 // vertically center child text
                                 align_items: AlignItems::Center,
-                                position: UiRect::default(),
                                 ..default()
                             },
                             background_color: NORMAL_BUTTON.into(),
@@ -514,7 +515,7 @@ pub fn hud_wall_setup(
         .spawn((
             NodeBundle {
                 style: Style {
-                    // size: Size::height(Val::Percent(42.)),
+                    // height: Val::Percent(42.),
                     flex_direction: FlexDirection::Row,
                     ..default()
                 },
@@ -530,14 +531,14 @@ pub fn hud_wall_setup(
                     .spawn((
                         ButtonBundle {
                             style: Style {
-                                size: Size::new(Val::Px(150.0), Val::Px(65.0)),
+                                width: Val::Px(150.0),
+                                height: Val::Px(65.0),
                                 // center button
                                 margin: UiRect::all(Val::Auto),
                                 // horizontally center child text
                                 justify_content: JustifyContent::Center,
                                 // vertically center child text
                                 align_items: AlignItems::Center,
-                                position: UiRect::default(),
                                 ..default()
                             },
                             background_color: NORMAL_BUTTON.into(),
@@ -570,10 +571,10 @@ pub fn hud_wall_setup(
             NodeBundle {
                 // image: character_sheet_resources.base_full_scroll.clone().into(),
                 style: Style {
-                    size: Size::height(Val::Percent(100.)),
+                    height: Val::Percent(100.),
                     flex_direction: FlexDirection::Column,
                     flex_shrink: 0.,
-                    position: UiRect::bottom(Val::Percent(100.)),
+                    bottom: Val::Percent(100.),
                     ..default()
                 },
                 visibility: Visibility::Hidden,
@@ -589,7 +590,7 @@ pub fn hud_wall_setup(
                 .spawn((
                     NodeBundle {
                         style: Style {
-                            size: Size::height(Val::Percent(20.)),
+                            height: Val::Percent(20.),
                             flex_direction: FlexDirection::Row,
                             ..default()
                         },
@@ -603,7 +604,7 @@ pub fn hud_wall_setup(
                         .spawn((
                             NodeBundle {
                                 style: Style {
-                                    size: Size::width(Val::Percent(30.)),
+                                    width: Val::Percent(30.),
                                     justify_content: JustifyContent::Center,
                                     ..default()
                                 },
@@ -618,7 +619,7 @@ pub fn hud_wall_setup(
                         .spawn((
                             NodeBundle {
                                 style: Style {
-                                    size: Size::width(Val::Percent(70.)),
+                                    width: Val::Percent(70.),
                                     flex_direction: FlexDirection::Column,
                                     ..default()
                                 },
@@ -633,7 +634,7 @@ pub fn hud_wall_setup(
                                 .spawn((
                                     NodeBundle {
                                         style: Style {
-                                            size: Size::height(Val::Percent(50.)),
+                                            height: Val::Percent(50.),
                                             flex_direction: FlexDirection::Row,
                                             ..default()
                                         },
@@ -647,7 +648,7 @@ pub fn hud_wall_setup(
                                 .spawn((
                                     NodeBundle {
                                         style: Style {
-                                            size: Size::height(Val::Percent(50.)),
+                                            height: Val::Percent(50.),
                                             flex_direction: FlexDirection::Row,
                                             ..default()
                                         },
@@ -664,7 +665,7 @@ pub fn hud_wall_setup(
                 .spawn((
                     NodeBundle {
                         style: Style {
-                            size: Size::height(Val::Percent(40.)),
+                            height: Val::Percent(40.),
                             flex_direction: FlexDirection::Row,
                             ..default()
                         },
@@ -679,7 +680,7 @@ pub fn hud_wall_setup(
                         .spawn((
                             NodeBundle {
                                 style: Style {
-                                    size: Size::width(Val::Percent(60.)),
+                                    width: Val::Percent(60.),
                                     flex_direction: FlexDirection::Column,
                                     justify_content: JustifyContent::Center,
                                     ..default()
@@ -704,7 +705,7 @@ pub fn hud_wall_setup(
                         .spawn((
                             NodeBundle {
                                 style: Style {
-                                    size: Size::width(Val::Percent(40.)),
+                                    width: Val::Percent(40.),
                                     justify_content: JustifyContent::Center,
                                     ..default()
                                 },
@@ -723,7 +724,8 @@ pub fn hud_wall_setup(
                                             .clone()
                                             .into(),
                                         style: Style {
-                                            size: Size::all(Val::Px(100.)),
+                                            width: Val::Px(100.),
+                                            height: Val::Px(100.),
                                             align_self: AlignSelf::Center,
                                             justify_content: JustifyContent::Center,
                                             ..default()
@@ -741,7 +743,7 @@ pub fn hud_wall_setup(
                 .spawn((
                     NodeBundle {
                         style: Style {
-                            size: Size::height(Val::Percent(40.)),
+                            height: Val::Percent(40.),
                             flex_direction: FlexDirection::Column,
                             // align_content: AlignContent::SpaceAround,
                             justify_content: JustifyContent::Center,
@@ -761,10 +763,8 @@ pub fn hud_wall_setup(
             //             style: Style {
             //                 flex_direction: FlexDirection::Column,
             //                 // flex_wrap: FlexWrap::NoWrap,
-            //                 size: Size {
-            //                     height: Val::Percent(100.),
-            //                     width: Val::Percent(100.),
-            //                 },
+            //                 height: Val::Percent(100.),
+            //                 width: Val::Percent(100.),
             //                 ..default()
             //             },
             //             ..default()
@@ -787,10 +787,8 @@ pub fn hud_wall_setup(
             //                     .clone()
             //                     .into(),
             //                 style: Style {
-            //                     size: Size {
-            //                         width: Val::Px(500.),
-            //                         height: Val::Percent(100.),
-            //                     },
+            //                     width: Val::Px(500.),
+            //                     height: Val::Percent(100.),
             //                     ..default()
             //                 },
             //                 ..default()
@@ -833,7 +831,7 @@ pub fn hud_wall_setup(
                 ImageBundle {
                     image: combat_wall_resources.base_combat_wall.clone().into(),
                     style: Style {
-                        size: Size::width(Val::Percent(36.)),
+                        width: Val::Percent(HUD_WALL_WIDTH),
                         flex_direction: FlexDirection::Column,
                         ..default()
                     },
@@ -862,7 +860,7 @@ pub fn hud_wall_setup(
                             style: Style {
                                 flex_shrink: 0.,
                                 flex_direction: FlexDirection::Column,
-                                size: Size::height(Val::Percent(100.)),
+                                height: Val::Percent(100.),
 
                                 ..default()
                             },
@@ -876,19 +874,7 @@ pub fn hud_wall_setup(
                             .spawn((
                                 NodeBundle {
                                     // background_color: Color::GRAY.into(),
-                                    style: Style {
-                                        flex_shrink: 0.,
-                                        flex_direction: FlexDirection::Column,
-                                        size: Size::new(Val::Percent(100.), Val::Percent(44.8)),
-                                        // gap between the two rows
-                                        gap: Size::height(Val::Percent(14.5)),
-                                        position: UiRect {
-                                            left: Val::Percent(16.8),
-                                            top: Val::Percent(4.2),
-                                            ..default()
-                                        },
-                                        ..default()
-                                    },
+                                    style: ALLIES_SHEET_STYLE,
                                     ..default()
                                 },
                                 Name::new("Allies' Scroll"),
@@ -897,14 +883,7 @@ pub fn hud_wall_setup(
                                 parent
                                     .spawn((
                                         NodeBundle {
-                                            style: Style {
-                                                flex_shrink: 0.,
-                                                flex_direction: FlexDirection::Row,
-                                                size: Size::height(Val::Percent(50.)),
-                                                // gap between the three scrolls
-                                                gap: Size::width(Val::Percent(2.7)),
-                                                ..default()
-                                            },
+                                            style: ROW_SHEETS_STYLE,
                                             ..default()
                                         },
                                         Name::new("First Row of Scrolls"),
@@ -925,6 +904,7 @@ pub fn hud_wall_setup(
                                                     } else {
                                                         Visibility::Hidden
                                                     },
+                                                    style: MINI_CHARACTER_SHEET_STYLE,
                                                     ..default()
                                                 },
                                                 Name::new(format!("Ally's Scroll {}", i)),
@@ -937,14 +917,7 @@ pub fn hud_wall_setup(
                                 parent
                                     .spawn((
                                         NodeBundle {
-                                            style: Style {
-                                                flex_shrink: 0.,
-                                                flex_direction: FlexDirection::Row,
-                                                size: Size::height(Val::Percent(50.)),
-                                                // gap between the three scrolls
-                                                gap: Size::width(Val::Percent(2.7)),
-                                                ..default()
-                                            },
+                                            style: ROW_SHEETS_STYLE,
                                             ..default()
                                         },
                                         Name::new("Second Row of Scrolls"),
@@ -965,6 +938,7 @@ pub fn hud_wall_setup(
                                                     } else {
                                                         Visibility::Hidden
                                                     },
+                                                    style: MINI_CHARACTER_SHEET_STYLE,
                                                     ..default()
                                                 },
                                                 Name::new(format!("Ally's Scroll {}", i)),
@@ -980,12 +954,9 @@ pub fn hud_wall_setup(
                                 image: combat_wall_resources.pack_of_scroll.clone().into(),
                                 style: Style {
                                     flex_shrink: 0.,
-                                    size: Size::width(Val::Percent(17.)),
-                                    position: UiRect {
-                                        left: Val::Percent(54.),
-                                        top: Val::Percent(31.8),
-                                        ..default()
-                                    },
+                                    width: Val::Percent(17.),
+                                    left: Val::Percent(54.),
+                                    top: Val::Percent(26.6),
                                     ..default()
                                 },
                                 ..default()
@@ -1003,12 +974,9 @@ pub fn hud_wall_setup(
                                 image: combat_log_resources.ladder.clone().into(),
                                 style: Style {
                                     flex_shrink: 0.,
-                                    size: Size::width(Val::Percent(32.)),
-                                    position: UiRect {
-                                        left: Val::Percent(9.7),
-                                        top: Val::Percent(25.7),
-                                        ..default()
-                                    },
+                                    width: Val::Percent(32.),
+                                    left: Val::Percent(9.7),
+                                    top: Val::Percent(20.5),
                                     ..default()
                                 },
                                 ..default()
@@ -1025,7 +993,7 @@ pub fn hud_wall_setup(
                         //             style: Style {
                         //                 flex_shrink: 0.,
                         //                 flex_direction: FlexDirection::Row,
-                        //                 // size: Size::height(Val::Percent(55.2)),
+                        //                 //  height: Val::Percent(55.2),
                         //                 ..default()
                         //             },
                         //             ..default()
@@ -1038,7 +1006,7 @@ pub fn hud_wall_setup(
                         //         //     .spawn((
                         //         //         NodeBundle {
                         //         //             style: Style {
-                        //         //                 size: Size::width(Val::Percent(50.)),
+                        //         //                 width: Val::Percent(50.),
                         //         //                 ..default()
                         //         //             },
                         //         //             ..default()
@@ -1051,7 +1019,7 @@ pub fn hud_wall_setup(
                         //         //     .spawn((
                         //         //         NodeBundle {
                         //         //             style: Style {
-                        //         //                 size: Size::width(Val::Percent(50.)),
+                        //         //                  width: Val::Percent(50.),
                         //         //                 ..default()
                         //         //             },
                         //         //             ..default()
@@ -1078,7 +1046,7 @@ pub fn global_ui_setup(mut commands: Commands, asset_server: Res<AssetServer>) {
         .spawn((
             NodeBundle {
                 style: Style {
-                    size: Size::width(Val::Percent(100.0)),
+                    width: Val::Percent(100.0),
                     justify_content: JustifyContent::SpaceBetween,
                     flex_direction: FlexDirection::Row,
                     ..default()
@@ -1097,7 +1065,7 @@ pub fn global_ui_setup(mut commands: Commands, asset_server: Res<AssetServer>) {
                 .spawn((
                     NodeBundle {
                         style: Style {
-                            size: Size::width(Val::Percent(56.)),
+                            width: Val::Percent(FIGHTING_HALL_WIDTH),
                             flex_direction: FlexDirection::Column,
                             ..default()
                         },
@@ -1112,12 +1080,10 @@ pub fn global_ui_setup(mut commands: Commands, asset_server: Res<AssetServer>) {
                         .spawn((
                             ButtonBundle {
                                 style: Style {
-                                    size: Size::new(Val::Px(200.0), Val::Px(65.0)),
+                                    width: Val::Px(200.0),
+                                    height: Val::Px(65.0),
                                     margin: UiRect::all(Val::Auto),
-                                    position: UiRect {
-                                        top: Val::Percent(5.),
-                                        ..default()
-                                    },
+                                    top: Val::Percent(5.),
                                     justify_content: JustifyContent::Center,
                                     align_items: AlignItems::Center,
                                     ..default()
@@ -1140,13 +1106,9 @@ pub fn global_ui_setup(mut commands: Commands, asset_server: Res<AssetServer>) {
                         .spawn((
                             NodeBundle {
                                 style: Style {
-                                    position: UiRect {
-                                        top: Val::Percent(5.),
-                                        ..default()
-                                    },
+                                    top: Val::Percent(5.),
                                     flex_direction: FlexDirection::Column,
                                     flex_grow: 1.0,
-                                    max_size: Size::UNDEFINED,
                                     ..default()
                                 },
                                 ..default()
@@ -1188,7 +1150,8 @@ pub fn global_ui_setup(mut commands: Commands, asset_server: Res<AssetServer>) {
                                 )
                                 .with_style(Style {
                                     flex_shrink: 0.,
-                                    size: Size::new(Val::Undefined, Val::Px(20.)),
+                                    width: Val::Px(0.),
+                                    height: Val::Px(20.),
                                     margin: UiRect {
                                         left: Val::Auto,
                                         right: Val::Auto,
@@ -1214,7 +1177,7 @@ pub fn global_ui_setup(mut commands: Commands, asset_server: Res<AssetServer>) {
                 .spawn((
                     NodeBundle {
                         style: Style {
-                            size: Size::width(Val::Percent(8.)),
+                            width: Val::Percent(INITIATIVE_BAR_WIDTH),
                             ..default()
                         },
                         background_color: Color::OLIVE.into(),