+++ src/ui/log_cave.rs
@@ -10,7 +10,7 @@ use bevy::{
 
 use crate::{
     combat::GameState,
-    constants::ui::style::*,
+    constants::ui::{style::*, HUD_WALL_WIDTH},
     ui::{
         combat_panel::{Ladder, UIScene},
         player_interaction::ScrollingList,
@@ -70,7 +70,8 @@ pub fn cave_ladder(
 ) {
     if let Ok(interaction) = ladder_query.get_single() {
         match interaction {
-            Interaction::Clicked => match game_state.0.clone() {
+            // TOTEST: `.clone()` needed ?
+            Interaction::Pressed => match game_state.get() {
                 GameState::CombatWall => {
                     next_state.set(GameState::LogCave);
                 }
@@ -120,7 +121,7 @@ pub fn setup(
                 ImageBundle {
                     image: combat_log_resources.base_log_cave.clone().into(),
                     style: Style {
-                        size: Size::width(Val::Percent(36.)),
+                        width: Val::Percent(HUD_WALL_WIDTH),
                         flex_direction: FlexDirection::Column,
                         ..default()
                     },
@@ -142,8 +143,8 @@ pub fn setup(
                             // it could be this linethat break the scrolling
                             flex_shrink: 0.,
                             // NOT QUITE RIGHT
-                            size: Size::width(Val::Percent(27.5)),
-                            position: UiRect::left(Val::Percent(7.3)), // -0.5
+                            width: Val::Percent(27.5),
+                            left: Val::Percent(7.3), // -0.5
                             ..default()
                         },
                         ..default()
@@ -159,10 +160,11 @@ pub fn setup(
                         NodeBundle {
                             style: Style {
                                 // flex_shrink: 0.,
-                                size: Size::new(Val::Percent(82.), Val::Percent(100.)),
+                                width: Val::Percent(82.),
+                                height: Val::Percent(100.),
                                 flex_direction: FlexDirection::Column,
                                 align_self: AlignSelf::Center,
-                                overflow: Overflow::Hidden,
+                                overflow: Overflow::clip_y(),
                                 ..default()
                             },
                             // background_color: Color::GRAY.into(),
@@ -194,7 +196,8 @@ pub fn setup(
                                     )
                                     .with_style(Style {
                                         flex_wrap: FlexWrap::Wrap,
-                                        size: Size::AUTO,
+                                        width: Val::Auto,
+                                        height: Val::Auto,
                                         // margin: UiRect {
                                         //     left: Val::Auto,
                                         //     right: Val::Auto,
@@ -216,7 +219,8 @@ pub fn setup(
                                     )
                                     .with_style(Style {
                                         flex_wrap: FlexWrap::Wrap,
-                                        size: Size::AUTO,
+                                        width: Val::Auto,
+                                        height: Val::Auto,
                                         // margin: UiRect {
                                         //     left: Val::Auto,
                                         //     right: Val::Auto,
@@ -238,7 +242,8 @@ pub fn setup(
                                     )
                                     .with_style(Style {
                                         flex_wrap: FlexWrap::Wrap,
-                                        size: Size::AUTO,
+                                        width: Val::Auto,
+                                        height: Val::Auto,
                                         // margin: UiRect {
                                         //     left: Val::Auto,
                                         //     right: Val::Auto,
@@ -260,7 +265,8 @@ pub fn setup(
                                     )
                                     .with_style(Style {
                                         flex_wrap: FlexWrap::Wrap,
-                                        size: Size::AUTO,
+                                        width: Val::Auto,
+                                        height: Val::Auto,
                                         ..default()
                                     }),
                                     Name::new("----"),